refactor orchestrator
